Remove BraintreeFragment.
