--- browser-switch/src/test/java/com/braintreepayments/browserswitch/test/TestBrowserSwitchFragment.java
@@ -1,25 +1,0 @@
-package com.braintreepayments.browserswitch.test;
-
-import android.net.Uri;
-
-import com.braintreepayments.browserswitch.BrowserSwitchFragment;
-import com.braintreepayments.browserswitch.BrowserSwitchResult;
-
-import androidx.annotation.Nullable;
-
-public class TestBrowserSwitchFragment extends BrowserSwitchFragment {
-
-    public boolean onBrowserSwitchResultCalled = false;
-    public int requestCode;
-    public BrowserSwitchResult result;
-    public Uri returnUri;
-
-    @Override
-    public void onBrowserSwitchResult(int requestCode, BrowserSwitchResult result,
-                                      @Nullable Uri returnUri) {
-        onBrowserSwitchResultCalled = true;
-        this.requestCode = requestCode;
-        this.result = result;
-        this.returnUri = returnUri;
-    }
-}
